Update React Native and its dependencies
React Native and its related packages have been upgraded. This includes several packages provided by '@react-native-community', '@react-native', and '@types/react-native'. Version 12.3.6 of the '@react-native-community/cli' was introduced and the primary React Native dependency was updated to version 0.73.5. This upgrade ensures the implementation of the latest features, updates, and security patches in the codebase.